requirements.txt
discord.py
langchain
python-dotenv
